test: share game session auth integration support
