+++ services/game-session-service/src/test/java/integration/net/firedevops/firemud/gamesession/GameSessionLoginIntegrationTest.java
@@ -10,9 +10,7 @@ import static org.mockito.Mockito.when;
 
 import java.net.URI;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
@@ -22,8 +20,8 @@ import net.firedevops.firemud.gamesession.dto.CommandEnqueueResult;
 import net.firedevops.firemud.gamesession.entity.GameInstance;
 import net.firedevops.firemud.gamesession.repository.GameInstanceRepository;
 import net.firedevops.firemud.gamesession.service.CommandService;
-import net.firedevops.firemud.gamesession.service.SessionContext;
 import net.firedevops.firemud.gamesession.service.SessionContextService;
+import net.firedevops.firemud.gamesession.testsupport.InMemorySessionContextTestConfiguration;
 import net.firedevops.firemud.test.NoGrpcServerTestConfiguration;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -31,11 +29,8 @@ import org.mockito.ArgumentCaptor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
-import org.springframework.context.annotation.Primary;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.web.socket.TextMessage;
 import org.springframework.web.socket.WebSocketHttpHeaders;
@@ -54,10 +49,7 @@ import org.springframework.web.socket.handler.TextWebSocketHandler;
       "spring.application.name=game-session-service",
       "spring.grpc.server.port=0"
     })
-@Import({
-  NoGrpcServerTestConfiguration.class,
-  GameSessionLoginIntegrationTest.InMemorySessionContextConfiguration.class
-})
+@Import({NoGrpcServerTestConfiguration.class, InMemorySessionContextTestConfiguration.class})
 class GameSessionLoginIntegrationTest {
   @LocalServerPort private int port;
 
@@ -120,56 +112,5 @@ class GameSessionLoginIntegrationTest {
     verify(accountClient)
         .authenticate(tenantCaptor.capture(), eq("demo@example.com"), eq("swordfish"), eq(""));
     assertThat(tenantCaptor.getValue()).isEqualTo("42");
-  }
-
-  @TestConfiguration(proxyBeanMethods = false)
-  static class InMemorySessionContextConfiguration {
-    @Bean
-    @Primary
-    SessionContextService sessionContextService() {
-      return new InMemorySessionContextService();
-    }
-
-    private static final class InMemorySessionContextService implements SessionContextService {
-      private final Map<Long, SessionContext> sessionMap = new ConcurrentHashMap<>();
-      private final Map<String, SessionContext> identityMap = new ConcurrentHashMap<>();
-
-      @Override
-      public void save(SessionContext context) {
-        sessionMap.put(context.sessionId(), context);
-        identityMap.put(identityKey(context), context);
-      }
-
-      @Override
-      public Optional<SessionContext> findByTenantAndSessionId(long tenantId, long sessionId) {
-        SessionContext context = sessionMap.get(sessionId);
-        if (context == null || context.tenantId() != tenantId) {
-          return Optional.empty();
-        }
-        return Optional.of(context);
-      }
-
-      @Override
-      public Optional<SessionContext> findByAccountAndCharacter(
-          long tenantId, long accountId, long characterId) {
-        return Optional.ofNullable(identityMap.get(identityKey(tenantId, accountId, characterId)));
-      }
-
-      @Override
-      public void deleteBySessionId(long tenantId, long sessionId) {
-        SessionContext removed = sessionMap.remove(sessionId);
-        if (removed != null) {
-          identityMap.remove(identityKey(removed));
-        }
-      }
-
-      private String identityKey(SessionContext context) {
-        return identityKey(context.tenantId(), context.accountId(), context.characterId());
-      }
-
-      private String identityKey(long tenantId, long accountId, long characterId) {
-        return tenantId + ":" + accountId + ":" + characterId;
-      }
-    }
   }
 }
