+++ services/game-session-service/src/test/java/integration/net/firedevops/firemud/gamesession/testsupport/InMemorySessionContextTestConfiguration.java
@@ -1,0 +1,62 @@
+package net.firedevops.firemud.gamesession.testsupport;
+
+import java.util.Map;
+import java.util.Optional;
+import java.util.concurrent.ConcurrentHashMap;
+import net.firedevops.firemud.gamesession.service.SessionContext;
+import net.firedevops.firemud.gamesession.service.SessionContextService;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Primary;
+
+@TestConfiguration(proxyBeanMethods = false)
+public class InMemorySessionContextTestConfiguration {
+
+  @Bean
+  @Primary
+  SessionContextService sessionContextService() {
+    return new InMemorySessionContextService();
+  }
+
+  private static final class InMemorySessionContextService implements SessionContextService {
+    private final Map<Long, SessionContext> sessionMap = new ConcurrentHashMap<>();
+    private final Map<String, SessionContext> identityMap = new ConcurrentHashMap<>();
+
+    @Override
+    public void save(SessionContext context) {
+      sessionMap.put(context.sessionId(), context);
+      identityMap.put(identityKey(context), context);
+    }
+
+    @Override
+    public Optional<SessionContext> findByTenantAndSessionId(long tenantId, long sessionId) {
+      SessionContext context = sessionMap.get(sessionId);
+      if (context == null || context.tenantId() != tenantId) {
+        return Optional.empty();
+      }
+      return Optional.of(context);
+    }
+
+    @Override
+    public Optional<SessionContext> findByAccountAndCharacter(
+        long tenantId, long accountId, long characterId) {
+      return Optional.ofNullable(identityMap.get(identityKey(tenantId, accountId, characterId)));
+    }
+
+    @Override
+    public void deleteBySessionId(long tenantId, long sessionId) {
+      SessionContext removed = sessionMap.remove(sessionId);
+      if (removed != null) {
+        identityMap.remove(identityKey(removed));
+      }
+    }
+
+    private String identityKey(SessionContext context) {
+      return identityKey(context.tenantId(), context.accountId(), context.characterId());
+    }
+
+    private String identityKey(long tenantId, long accountId, long characterId) {
+      return tenantId + ":" + accountId + ":" + characterId;
+    }
+  }
+}
